+++ library/unified/server/source/partial_png_compressor_wrapper.cpp
@@ -28,13 +28,13 @@ void sirius::library::unified::partialpng::compressor::after_process_callback(in
 		_compressor->after_video_compressing_callback(index, compressed, size, before_compress_timestamp, after_compress_timestamp);
 }
 
-void sirius::library::unified::partialpng::compressor::after_process_callback(int32_t count, int32_t * x, int32_t * y, int32_t * width, int32_t * height, uint8_t ** compressed, int32_t * size, long long before_compress_timestamp, long long after_compress_timestamp)
+void sirius::library::unified::partialpng::compressor::after_process_callback(int32_t count, int16_t * x, int16_t * y, int16_t * width, int16_t * height, uint8_t ** compressed, int32_t * size, long long before_compress_timestamp, long long after_compress_timestamp)
 {
 	if (_compressor)
 		_compressor->after_video_compressing_callback(count, x, y, width, height, compressed, size, before_compress_timestamp, after_compress_timestamp);
 }
 
-void sirius::library::unified::partialpng::compressor::after_process_callback(int32_t x, int32_t y, int32_t width, int32_t height, uint8_t * compressed, int32_t size, long long before_compress_timestamp, long long after_compress_timestamp)
+void sirius::library::unified::partialpng::compressor::after_process_callback(int16_t x, int16_t y, int16_t width, int16_t height, uint8_t * compressed, int32_t size, long long before_compress_timestamp, long long after_compress_timestamp)
 {
 	if (_compressor)
 		_compressor->after_video_compressing_callback(x, y, width, height, compressed, size, before_compress_timestamp, after_compress_timestamp);
