+++ library/video/transform/codec/base/source/sirius_video_compressor.cpp
@@ -90,12 +90,12 @@ void sirius::library::video::transform::codec::compressor::after_process_callbac
 
 }
 
-void sirius::library::video::transform::codec::compressor::after_process_callback(int32_t count, int32_t * x, int32_t * y, int32_t * width, int32_t * height, uint8_t ** compressed, int32_t * size, long long before_compress_timestamp, long long after_compress_timestamp)
+void sirius::library::video::transform::codec::compressor::after_process_callback(int32_t count, int16_t * x, int16_t * y, int16_t * width, int16_t * height, uint8_t ** compressed, int32_t * size, long long before_compress_timestamp, long long after_compress_timestamp)
 {
 
 }
 
-void sirius::library::video::transform::codec::compressor::after_process_callback(int32_t x, int32_t y, int32_t width, int32_t height, uint8_t * compressed, int32_t size, long long before_compress_timestamp, long long after_compress_timestamp)
+void sirius::library::video::transform::codec::compressor::after_process_callback(int16_t x, int16_t y, int16_t width, int16_t height, uint8_t * compressed, int32_t size, long long before_compress_timestamp, long long after_compress_timestamp)
 {
 
 }
